refactor(review): 拆分 calendar_event_utils 与 course.dart 上帝文件
calendar_event_utils.dart (749行) → 3 职责模块 + barrel：
- calendar_event_models.dart：模型类（StructuredLocation/ResolvedLocation/ExportPayload/EventPayload）
- calendar_location_mapper.dart：地点解析 + 静态地理数据（~620行大头）
- calendar_event_identity.dart：稳定 UID 生成
- calendar_event_utils.dart：re-export barrel，外部 import 零改动

course.dart (622行) → course.dart + schedule_config.dart：
- course.dart：Course/WeekType/DateTimeExtension + 顶层 kDefaultTotalWeeks + export barrel
- schedule_config.dart：TimeSlot/ScheduleConfig（唯一定义），单向 import course.dart
- 避免循环依赖：常量提升为顶层，ScheduleConfig.kDefaultTotalWeeks 两处调用改 kDefaultTotalWeeks
